feat: export types for consuming
